requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.4
sqlalchemy==2.0.31
psycopg2-binary
celery==5.4.0
redis==5.0.7
google-generativeai==0.7.2
python-multipart==0.0.9
httpx==0.27.0
email-validator==2.2.0